Bare Bones Scene Transitioning & Tutorials
Currently plays chopping scene, then slicing scene.
Tutorial video will play at the start of the scene and will replay until player clicks the close button.
In the chopping scene, once the player has chopped the onion, the tomato will be enabled (both assets are enabled in the editor to allow the script to find the tomato object. It will be disabled automatically on start).
Scenes currently transition in a loop (chopping->slicing->chopping) so technically any scene can be played first.