Added Cutting "Zones"
Added cutting zones so that play has to click and drag from one edge of ingredient to the other to make the cut.

Ingredient Collider reference needs to be manually set in cutting knife inspector depending on which ingredient (tofu/chicken) is active in scene.

Since chicken edge is uneven, clicking and dragging in the middle of the chicken might result in a cut being made depending on if the colliders are hit.