Replace set trigger with crossfade
This fixes issues with the opacity of the dialogue box tooltips for spices. The duration of the fade is constant, i.e. the change in opacity is not constant, but changing that would require knowing the max alpha and minimum alpha or other shenanigans that likely aren't worth the effort; the speed of the animation makes this a non-issue, for all intents and purposes