Made WebGL downloading work (but not smoothly)
The user needs to click the button and then needs to click anywhere on the canvas, including the button, to download the file. Doing this does not bring up a save file dialog. The console says "ClientWaitSync must return TIMEOUT_EXPIRED until control has returned to the user agent's main loop." on every button click.

Next steps are to figure out what the DownloadFile imported from the plugins folder with DllImport internal actually is, so that it can perhaps be reverse engineered to not cause problems(?).