Made a canvas for showing/downloading the recipe
Displaying pdfs is an Ordeal™ so instead, I converted the first (and only) page of the test pdf into a jpg to use as a preview.

Actually inserting this canvas into the dialog sequence as it is wouldn't be trivial, and since it's liable to undergo drastic changes soon, the canvas is not actually implemented. Its prefab should only need to be dropped into the scene for it to work, though.