Made hand not follow mouse off-screen
This prevents the hand from getting stuck behind the cursor-only zone in the top right of spice selection.

I'm not sure I like the cursor-only zone, but the hand will work identically if we just turn that off, especially since there's an inspector bool to toggle the hand following the mouse off screen.