Further improvements to no offscreen cursor following
Same as before; works identically to the original movement in typical circumstances.

The mouse can *STILL* get stuck behind the cursor only zone if you move the mouse fast enough, thus getting it far enough out before it can stop following the cursor. This should be uncommon in normal play.

It'd be ideal if the cursor only zone only blocked collisions coming from the left and right, though, to prevent even this. If there's a simple way to do that, all of this complicated math/mental gymnasticing could have been avoided, and that makes me an unhappy gamer.