More fiddling with pan friction
Was prepping for changing MoveWithMousePos to use AddForce instead of MovePosition, but found that changing the pan to dynamic instead of kinematic makes the friction behave *ENTIRELY* differently.

And also, realized this was supposed to be a pot and not a pan the whole time. Whoops. A bunch of this work can still be reused, at least!