Very belated web build update
Somehow, in all the excitement of later development, we never updated the web build. My theory is every milestone we worked towards from then on needed a standalone build, but not a web build, causing the latter to fall by the wayside. I only noticed by double checking some links in my career materials, one of which pointed to the web build.

There may be unfinished content that's improperly accessible, or debug elements that haven't been stripped out in this build. If there are, I reckon that's fine. It's not entirely trivial to sniff those kinds of things out after a few years away from a project, anyway.